% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloc}
\alias{bloc}
\title{bloc}
\format{
Data frame with 07 columns  and 51,397 rows
\describe{
\item{STRUCTURE}{établissement.}
\item{AGE}{age en année.}
\item{SEXE}{sexe.}
\item{DATE_INTERVENTION}{date intervention.}
\item{MEDECIN}{médecin.}
\item{DIAGNOSTIC}{diagnostic.}
\item{TYPE}{type.}
}
}
\source{
DSP Djelfa
}
\usage{
bloc
}
\description{
Look up bloc.
}
\examples{
  bloc
}
\keyword{datasets}
