% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget}
\alias{budget}
\title{budget section 2 dépence}
\format{
Data frame with 05 columns  and 420rows
\describe{
\item{SECTION}{section.}
\item{TITRE}{titre.}
\item{CHAPITRE}{chapitre.}
\item{MONTANT}{montant.}
\item{STR}{structure sanitaire.}
}
}
\source{
DSP Djelfa
}
\usage{
budget
}
\description{
Look up budget.
}
\examples{
  budget
}
\keyword{datasets}
