% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19}
\alias{covid19}
\title{covid19}
\format{
Data frame with 05 columns  and 420rows
\describe{
\item{SECTION}{section.}
\item{TITRE}{titre.}
\item{CHAPITRE}{chapitre.}
\item{MONTANT}{montant.}
\item{STR}{structure sanitaire.}
}
}
\source{
DSP Djelfa
}
\usage{
covid19
}
\description{
Look up covid19.
}
\examples{
  covid19
}
\keyword{datasets}
