% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.dsp-package.R
\docType{package}
\name{datasets.dsp-package}
\alias{datasets.dsp}
\alias{datasets.dsp-package}
\title{datasets.dsp: Provides the RDP datasets.dsp}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides the RDP datasets.dsp.
}
\author{
\strong{Maintainer}: Redha TIBA \email{tibaredha@yahoo.fr}

}
\keyword{internal}
