% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deceshosp}
\alias{deceshosp}
\title{mortalité hospitalière}
\format{
Data frame with 19 columns  and 20584 rows
\describe{
\item{DINS}{date déces.}
\item{DATENAISSANCE}{date de naissance.}
\item{SEX}{sexe.}
\item{Years}{age en année.}
\item{Days}{age en jours.}
\item{WILAYAR}{wilaya de résidence.}
\item{WILAYAS}{wilaya de naissance.}
\item{COMMUNER}{comuune de résidence.}
\item{COMMUNE}{commune de naissance.}
\item{LD}{lieux du déces.}
\item{structure}{établissement du déces.}
\item{service}{service du déces.}
\item{DUREEHOSPIT}{durée d'hospitalisation.}
\item{n_profession}{la profession.}
\item{CD}{cause du déces.}
\item{CHAP}{chapitre.}
\item{diag_cod}{code cim.}
\item{diag_nom}{dignostic.}
\item{MEDECINHOSPIT}{médecin.}
}
}
\source{
DSP Djelfa
}
\usage{
deceshosp
}
\description{
Look up deceshosp.
}
\examples{
  deceshosp
}
\keyword{datasets}
