% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_oph_2025}
\alias{df_oph_2025}
\title{df_oph_2025}
\format{
Data frame with 07 columns  and 26,568 rows
\describe{
\item{STRUCTURE}{établissement.}
\item{AGE}{age en année.}
\item{SEXE}{sexe.}
\item{DATE_INTERVENTION}{date intervention.}
\item{MEDECIN}{médecin.}
\item{DIAGNOSTIC}{diagnostic.}
\item{TYPE}{type.}
}
}
\source{
DSP Djelfa
}
\usage{
df_oph_2025
}
\description{
Look up df_oph_2025.
}
\examples{
  df_oph_2025
}
\keyword{datasets}
