% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rh}
\alias{rh}
\title{rh}
\format{
Data frame with 09 columns  and 9364 rows
\describe{
\item{wilaya}{wilaya.}
\item{structure}{structure.}
\item{sexe}{sexe.}
\item{datenaissance}{datenaissance.}
\item{grade}{grade.}
\item{specialité}{specialité.}
\item{nom}{nom.}
\item{prenom}{prenom.}
}
}
\source{
DSP Djelfa
}
\usage{
rh
}
\description{
Look up rh.
}
\examples{
  rh
}
\keyword{datasets}
