% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig.c}
\alias{sig.c}
\title{géométrie communes}
\format{
Data frame with 04 columns  and  1504 rows
\describe{
\item{CC_2}{code commune.}
\item{NAME_1}{nom wilaya.}
\item{NAME_2}{nom commune.}
\item{geometry}{geometry.}
}
}
\source{
DSP Djelfa
}
\usage{
sig.c
}
\description{
Look up sig.c.
}
\examples{
  sig.c
}
\keyword{datasets}
