% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig.w}
\alias{sig.w}
\title{géométrie wilayas}
\format{
Data frame with 03 columns  and  48 rows
\describe{
\item{CC_1}{code wilaya.}
\item{NAME_1}{nom wilaya.}
\item{geometry}{geometry.}
}
}
\source{
DSP Djelfa
}
\usage{
sig.w
}
\description{
Look up sig.w.
}
\examples{
  sig.w
}
\keyword{datasets}
