% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sscolaire}
\alias{sscolaire}
\title{sante scolaire}
\format{
Data frame with 07 columns  and 170 rows
\describe{
\item{etablissement}{etablissement.}
\item{PRE}{PRE.}
\item{1F}{1F.}
\item{2F}{2F.}
\item{3F}{3F.}
\item{4F}{4F.}
\item{5F}{5F.}
\item{1M}{1m.}
\item{2M}{2M.}
\item{3M}{3M.}
\item{4M}{4M.}
\item{1S}{1S.}
\item{2S}{2S.}
\item{3S}{3S.}
\item{UDS}{UDS.}
}
}
\source{
DSP Djelfa
}
\usage{
sscolaire
}
\description{
Look up sscolaire.
}
\examples{
  sscolaire
}
\keyword{datasets}
